More fixes related to Bug 41
Updated DirectoryViewControl to take misc used space calculation from TreeContext, which also calculates it, and already guarded against it wrapping.
The latter calculation contained an error, which has been fixed.
Furthermore, in case of a misc used space anomaly, updating size of free space item in volume tree so that total size of volume tree still matches that of the claimed volume size.