Bug fix: Let Quick Look panel close after selection changed
The initial Quick Look implementation assumed that the selection remained unchanged when the Quick Look panel was chosen (and even asserted that). That, however, was invalid. It is possible to change the selection with the Quick Look panel open. The result was that the panel could not close after a selection change.

This has been fixed by letting the DirectoryViewPreviewItem retain the path to the selected item and use that for all subsequent operations on the the item.