Refactored TreeBuilder to use NSDirectoryEnumerator
The usage of NSDirectoryEnumerator replaces the usage of the deprecated FSGetCatalogInfoBulk API.

The enumerator returns NSURL objects. So also refactored usage of other deprecated APIs to get path properties to use NSURL's resource values instead.

As the order of the enumeration cannot be controlled, there is one regression in functionality. The tree builder used to scan oldest directories first. This impacted where hard-linked items appeared. In Time Machine back-ups it showed up in the first back-up. This can no longer be guaranteed. The original logic could be retained by a shallow traversal of each directory. However, this presumably has a big impact on performance, as creation of the enumerator looks fairly expensive.

The code compiles, but no attempt was yet made to run it.