Minor fix to TreeBuilder
Only skipDescendants when the hard-linked item is a directory.
With this change, the scan results are similar to that created by the previous version of the app. The total number of files, folders and total size are the same. There are some minor changes in the files that are part of the tree. This is expected, given that both versions may include a different copy of each hard-linked file with multiple copies.