Changed internal component representation for FileItems
FileItem now stores the systemPath internally. It is converted to a friendly path where needed.
This is done for several reasons. One, this is how the path is obtained from NSURL. Two, it matches how localization is handled for the name of non-physical file items. Three, it actually simplifies the code.
In doing so, renamed the old "name" to "label", mainly to force review of all impacted code.
TreeWriter now also writes the system representation for all paths. Updated TreeReader to automagically convert old formats. Introduced reading of format version to do so efficiently (i.e. only convert where needed).