Updated deployment target settings
Set deployment target to OS X 10.9 (which according to info on public website is the required MacOS version; not sure anymore what the specific dependency is though).
Also updated SDK root to 10.12. I assume this was the effective setting anyway, as the SDK of the old setting could not even be found.