Show warning when logical size is larger than physical size
The logical file size can be larger than the physical size.
This can happen for HFS+ compressed files and also for sparse files on APFS.
When this happens a warning is now shown in the resulting DirectoryView window.
First, a warning is shown when the total scanned size is larger than the actual size.
If that's not the case, a warning is shown when the size of some files is larger than their actual size.