Increased size of text fields in recent scan table
These fields were quite small, making it quite likely that long folder paths were truncated after the welcome window was increased in size.
The fields are now big enough that this should not happen anymore in practise.
It's still theoretically possible, but it's not worth investing in a dynamic sizing solution for this corner case.