Updated display in status bar
When the tally measure is used and a single file is selected, its size is not shown.
Otherwise, when the tally measure is used, the "files" unit is also shown for clarity.

Also swapped the path and size fields.
The path is shown first, the size right-aligned at the right.
This is more logical when the size field can be empty and anyway looks better.

For consistency, moved static helper method to generate exact file size string from FileItem to DirectoryViewControl.