Animate overlay
When window is main and an overlay is active, vary its intensity. This helps to find the
selected items.

Also removed use of Core Graphics for drawing overlay. It is not needed now that only the
overlay image uses transparency and can be rendered on top of basic image. This improves
performance and simplifies the code.