Do not explicitly set size anymore for toolbar items
This addresses the following warning:
"NSToolbarItem.minSize and NSToolbarItem.maxSize methods are deprecated. Usage may result in clipping of items. It is recommended to let the system measure the item automatically using constraints."

So far, after some basic testing, this seems to work fine.
It appears there no need to replace this by coded constraints.